src/main/java/com/ensak/positions_app/customer/CustomerController.java
package com.ensak.positions_app.customer;

import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/customers")
public class CustomerController {
    private CustomerService customerService;

    public CustomerController(CustomerService customerService) {
        this.customerService = customerService;
    }

    @GetMapping
    public List<CustomerDto> search(){
        return customerService.search();
    }
    @GetMapping("/{id}")
    public CustomerDto read(@PathVariable int id){
        return customerService.read(id);
    }
}
